+++ packages/stepper/src/StepIcon/StepIcon.styles.ts
@@ -3,7 +3,7 @@ import { Theme } from '@leafygreen-ui/lib';
 import { palette } from '@leafygreen-ui/palette';
 import { spacing, transitionDuration } from '@leafygreen-ui/tokens';
 
-import { type StepState, StepStates } from '../Stepper';
+import { type StepState, StepStates } from '../types';
 
 export const baseStyles = css`
   box-sizing: content-box;
