+++ packages/stepper/src/StepIcon/StepIconGlyph.tsx
@@ -6,7 +6,7 @@ import EllipsisIcon from '@leafygreen-ui/icon/dist/Ellipsis';
 import { fontWeights } from '@leafygreen-ui/tokens';
 import { Overline } from '@leafygreen-ui/typography';
 
-import { StepStates } from '../Stepper';
+import { StepStates } from '../types';
 
 import { StepIconProps } from './StepIcon.types';
