+++ packages/stepper/src/StepLabel/StepLabel.tsx
@@ -6,7 +6,7 @@ import { useDarkMode } from '@leafygreen-ui/leafygreen-provider';
 import { Body } from '@leafygreen-ui/typography';
 
 import { stepLabelClassName } from '../constants';
-import { StepStates } from '../Stepper';
+import { StepStates } from '../types';
 
 import { getThemedStateColorStyles, multipleStyles } from './StepLabel.styles';
 import { StepLabelProps } from './StepLabel.types';
