+++ packages/stepper/src/types.ts
@@ -1,0 +1,9 @@
+export const StepStates = {
+  CompletedMultiple: 'completed-multiple',
+  Completed: 'completed',
+  Current: 'current',
+  Upcoming: 'upcoming',
+  UpcomingMultiple: 'upcoming-multiple',
+};
+
+export type StepState = (typeof StepStates)[keyof typeof StepStates];
